Variable sound values for player
The sound values of the player's movement states can variable from -1 to 1 so as to not keep them a constant value. This is to ensure that even when the player is walking they can still trigger mines they walk close to. The mines in addition have increased sound sensitivity when the player is closer to them and can activate at the tiniest sound